+++ Stripe/STPPaymentIntentParams.m
@@ -43,9 +43,15 @@
                        [NSString stringWithFormat:@"clientSecret = %@", (self.clientSecret.length > 0) ? @"<redacted>" : @""],
                        [NSString stringWithFormat:@"receiptEmail = %@", self.receiptEmail],
                        [NSString stringWithFormat:@"returnURL = %@", self.returnURL],
-                       [NSString stringWithFormat:@"saveSourceToCustomer = %@", (self.saveSourceToCustomer.boolValue) ? @"YES" : @"NO"],
+                       [NSString stringWithFormat:@"savePaymentMethod = %@", (self.savePaymentMethod.boolValue) ? @"YES" : @"NO"],
+
+                       // Source
                        [NSString stringWithFormat:@"sourceId = %@", self.sourceId],
                        [NSString stringWithFormat:@"sourceParams = %@", self.sourceParams],
+                       
+                       // PaymentMethod
+                       [NSString stringWithFormat:@"paymentMethodId = %@", self.paymentMethodId],
+                       [NSString stringWithFormat:@"paymentMethodParams = %@", self.paymentMethodParams],
 
                        // Additional params set by app
                        [NSString stringWithFormat:@"additionalAPIParameters = %@", self.additionalAPIParameters],
@@ -64,6 +70,14 @@
     self.returnURL = returnUrl;
 }
 
+- (NSNumber *)saveSourceToCustomer {
+    return self.savePaymentMethod;
+}
+
+- (void)setSaveSourceToCustomer:(NSNumber *)saveSourceToCustomer {
+    self.savePaymentMethod = saveSourceToCustomer;
+}
+
 #pragma mark - STPFormEncodable
 
 + (nullable NSString *)rootObjectName {
@@ -73,10 +87,12 @@
 + (nonnull NSDictionary *)propertyNamesToFormFieldNamesMapping {
     return @{
              NSStringFromSelector(@selector(clientSecret)): @"client_secret",
+             NSStringFromSelector(@selector(paymentMethodParams)): @"payment_method_data",
+             NSStringFromSelector(@selector(paymentMethodId)): @"payment_method",
              NSStringFromSelector(@selector(sourceParams)): @"source_data",
              NSStringFromSelector(@selector(sourceId)): @"source",
              NSStringFromSelector(@selector(receiptEmail)): @"receipt_email",
-             NSStringFromSelector(@selector(saveSourceToCustomer)): @"save_source_to_customer",
+             NSStringFromSelector(@selector(savePaymentMethod)): @"save_payment_method",
              NSStringFromSelector(@selector(returnURL)): @"return_url",
              };
 }